requirements.txt
django-allauth 
django-debug-toolbar 
django-bootstrap-v5
whitenoise 
celery